+++ src/renderer/screens/OpenGov/Tracks/index.tsx
@@ -7,11 +7,12 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCaretLeft, faInfo } from '@fortawesome/pro-solid-svg-icons';
 import { ContentWrapper, HeaderWrapper } from '@app/screens/Wrappers';
 import { DragClose } from '@/renderer/library/DragClose';
-import { OpenGovFooter, Scrollable, TrackGroup } from './Wrappers';
+import { OpenGovFooter, Scrollable } from '../Wrappers';
+import { TrackGroup } from './Wrappers';
 import { ButtonPrimaryInvert } from '@/renderer/kits/Buttons/ButtonPrimaryInvert';
 import { TrackRow } from './TrackRow';
 import type { HelpItemKey } from '@/renderer/contexts/common/Help/types';
-import type { TracksProps } from './types';
+import type { TracksProps } from '../types';
 
 export const Tracks = ({ setSection, chainId }: TracksProps) => {
   /// Context data.
